requirements_lock.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile --allow-unsafe --output-file=requirements_lock.txt requirements.in
#
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.12.13
    # via
    #   -r requirements.in
    #   aiohttp-cors
aiohttp-cors==0.8.1
    # via -r requirements.in
aiosignal==1.3.2
    # via aiohttp
attrs==25.3.0
    # via aiohttp
certifi==2025.4.26
    # via requests
charset-normalizer==3.4.2
    # via requests
frozenlist==1.7.0
    # via
    #   aiohttp
    #   aiosignal
grpcio==1.71.0
    # via
    #   -r requirements.in
    #   grpcio-health-checking
    #   grpcio-tools
grpcio-health-checking==1.71.0
    # via -r requirements.in
grpcio-tools==1.71.0
    # via -r requirements.in
idna==3.10
    # via
    #   requests
    #   yarl
iniconfig==2.1.0
    # via pytest
multidict==6.5.0
    # via
    #   aiohttp
    #   yarl
packaging==25.0
    # via pytest
pluggy==1.5.0
    # via pytest
propcache==0.3.2
    # via
    #   aiohttp
    #   yarl
protobuf==5.29.4
    # via
    #   grpcio-health-checking
    #   grpcio-tools
psycopg2-binary==2.9.10
    # via -r requirements.in
pytest==8.3.5
    # via -r requirements.in
requests==2.32.3
    # via -r requirements.in
urllib3==2.4.0
    # via requests
yarl==1.20.1
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
setuptools==80.2.0
    # via grpcio-tools
